# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server API for NodeList program",
        "title": "NodeList API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "/v1",
    "paths": {
        "/license/amount": {
            "get": {
                "description": "Get all license amount that have been sold, and all license amount that have been delegated",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "License"
                ],
                "summary": "Get all license amount and delegated license amount",
                "responses": {
                    "200": {
                        "description": "return the amount",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/license/amount/owner/{address}": {
            "get": {
                "description": "Get the license amount that the wallet address has purchased",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "License"
                ],
                "summary": "Get all license amount of the owner",
                "parameters": [
                    {
                        "type": "string",
                        "description": "owner address(an ethereum address with prefix '0x')",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return amount",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "request parameter error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/license/info/owner/{address}": {
            "get": {
                "description": "Query the license information owned by the owner through the wallet address, support paging",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "License"
                ],
                "summary": "Get the license information of the specified owner in pages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "owner address(an ethereum address with prefix '0x')",
                        "name": "address",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "paging start index (default 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "number of items to return per page(default 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return license info list successfully",
                        "schema": {
                            "$ref": "#/definitions/server.LicenseInfos"
                        }
                    },
                    "400": {
                        "description": "request parameter error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/license/price": {
            "get": {
                "description": "Get license price, include how many USDT and how many ETH",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "License"
                ],
                "summary": "Get license price",
                "responses": {
                    "200": {
                        "description": "return the price",
                        "schema": {
                            "$ref": "#/definitions/server.LicensePrice"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "server.LicenseInfo": {
            "type": "object",
            "properties": {
                "delegated": {
                    "type": "boolean"
                },
                "delegatedNode": {
                    "type": "string"
                },
                "initialReward": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                },
                "tokenID": {
                    "type": "string"
                },
                "totalReward": {
                    "type": "string"
                },
                "withdrawedReward": {
                    "type": "string"
                }
            }
        },
        "server.LicenseInfos": {
            "type": "object",
            "properties": {
                "infos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.LicenseInfo"
                    }
                }
            }
        },
        "server.LicensePrice": {
            "type": "object",
            "properties": {
                "eth": {
                    "description": "xxxETH/1License",
                    "type": "string"
                },
                "usdt": {
                    "description": "xxxUSDT/1License",
                    "type": "string"
                }
            }
        },
...
    }
}